Add A&AI actor and some operators
Added A&AI Actor, as well as operators for the "custom query" and
the "tenant" query (on which the custom query depends).

Issue-ID: POLICY-1625

+++ models-interactions/model-actors/actorServiceProvider/src/test/java/org/onap/policy/controlloop/actorserviceprovider/parameters/TopicParamsTest.java
@@ -36,7 +36,7 @@ public class TopicParamsTest {
     private static final String CONTAINER = "my-container";
     private static final String TARGET = "my-target";
     private static final String SOURCE = "my-source";
-    private static final long TIMEOUT = 10;
+    private static final int TIMEOUT = 10;
 
     private TopicParams params;
